verilog/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.2
image==1.5.27
importlib-metadata==0.18
more-itertools==7.0.0
numpy==1.16.4
packaging==19.0
pathlib2==2.3.3
Pillow==6.0.0
pkg-resources==0.0.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.1
